backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
yfinance>=0.2.40
pandas>=2.2.3
numpy>=1.26
scikit-learn>=1.4
